+++ DataLayer/Event/Purchase.php
@@ -37,7 +37,7 @@ class Purchase implements EventInterface
                 'transaction_id' => $order->getIncrementId(),
                 'affiliation' => $this->config->getStoreName(),
                 'currency' => $order->getOrderCurrencyCode(),
-                'value' => $this->priceFormatter->format((float)$order->getGrandTotal()),
+                'value' => $this->priceFormatter->format($this->getPurchaseValue($order)),
                 'tax' => $this->priceFormatter->format((float)$order->getTaxAmount()),
                 'shipping' => $this->priceFormatter->format((float)$order->getShippingAmount()),
                 'coupon' => $order->getCouponCode(),
@@ -54,5 +54,10 @@ class Purchase implements EventInterface
     {
         $this->order = $order;
         return $this;
+    }
+
+    private function getPurchaseValue(OrderInterface $order): float
+    {
+        return (float)$order->getGrandTotal() - (float)$order->getTaxAmount() - (float)$order->getShippingAmount();
     }
 }